Refactor iscsi initiator utils
At the moment, all the iSCSI initiator utils classes can be found
in the volumeutils modules. This is quite misleading, especially
as we are going to add iSCSI target server utils classes and
other initiator types.

This patch moves and renames the current volumeutils modules to:

storage
    initiator
        base_iscsi_utils.py
        iscsi_cli_utils.py
        iscsi_wmi_utils.py

The utilsfactory get_volumeutils method is left for backwards
compatibility with Nova.

Change-Id: I1bc6628dc842595de3fb6daa4f39ca330609cfa2