Fix method retrieving share capacity info
This method uses the 'run_and_check_output' wrapper for executing a
kernel32 lib function. The issue is that the 'kernel32_lib_func' flag
was omitted, for which reason the error handling is done improperly.

Namely, this function will return non-zero values in case of success,
which at the moment is interpreted as an error.

This patch fixes the issue by adding the missing flag.

Change-Id: I63c7323ac9a5a4bd850971b2de67012b78bdbfa5