Maintain livemigrationutils compatibility
Live migration related utils have been refactored and it requires
the planned VM to be created on the destination host beforehand.
In order to maintain compatibility with the current usage of the
mentioned utils, if there is no planned VM on the destination
host, it has to be created.

This will be removed once the usage has been updated.

Change-Id: I48dd108eab75ed150128ff6cf390a0fc4b7bb74e
Related: blueprint hyperv-fibre-channel