Creates baseutils module
Creates BaseUtils class, which contains the _get_wmi_obj and
_get_wmi_conn methods, commonly used in most of the utils
classes.

Caches the WMI connections, as they do not have to be unique.

Sets the WMI _conn object, the Msvm_VirtualSystemManagementService
as object attributes, as they are constantly used and there is no
need to instantiate them every time.

Partial-Bug: #1505196

Change-Id: Ieee7b9e587424f2a95aee85389adfaabf4032779