Add diskutils module
This patch adds a new module, used for disk related operations.

It adds a new method for retrieving disk unique id. We need this
method for os-brick, which returns the Unique id of the attached
volumes.

Note that for iSCSI/FC volumes, os-brick uses scsi_id to retrieve
the page 83 SCSI identifier, prepending the id type. We want to
remain consistent with the information provided by Linux os-brick
connectors so we'll do the same.

The method used for retrieving the disk number from a device name
has also been moved in this new module, previously residing in
vmutils as it will also be used by os-brick.

Also, the method used for rescanning disks has been moved here.

Change-Id: I0eac17c317a8d25da44b86feb58d6e8556f642eb