Fibre Channel: Fix invalid HBA adapter issues
The FC hbaapi library can report invalid HBAs, which it cannot
handle.

For example, in case of a VMWare based VM, it can report the
SCSI controller as a FC HBA.

We can avoid having issues because of this by simply ignoring
adapters that could not be processed, logging a meaningful message.

Change-Id: I46285521ceb60d05a4ccc50b1ccf692bd50d976c
Closes-Bug: #1539702