Fix getting attached VHD images info
In some environments, the 'VIRTUAL_DISK_ACCESS_GET_INFO' access
mask is not handled properly, for which reason we won't be able
to open attached VHD images.

This patch fixes this issue by explicitly requesting read only
access mode, also setting the "GetInfoOnly" flag. The only issue
with this approach is that it's not supported on Windows versions
prior to Windows Server 2012, but that's fine as we've dropped
support for Windows Server 2008 R2.

Note that the "VIRTUAL_DISK_ACCESS_NONE" may look a bit odd but this
access mask is required to be set when using the V2 open parameters
structure.

Change-Id: I7e5f088a78b5ec8f8695d5ba677f23496829fbc2