Caches Security Group Rule ACLs
In order to avoid querying for the mentioned WMI objects, adds
a cache for them, verifying the existence of the Security Group
Rules using the cache instead.

Adds the Weight field to the SecurityGroupRuleR2 in order for the
objects to have the same fields as the ACLs.

(cherry picked from commit b14a331637a43a7230d982683b4ab12f3eaf401c)

Partial-Bug: #1505196

Change-Id: I63d5984f3f9343e32e065f219567c4a821eb997b