Fix retry decorators
Some of our jobutils methods use loopingcall.RetryDecorator,
having a maximum number of retries. The issue with that is that
the implementation of this decorator uses a static counter for
the number of retries performed.

In other words, after the maximum number of retries is performed,
no more retries will be performed even if the method is called
in a different context.

This patch adds a new decorator, fixing this issue. It can also
accept expected error codes.

Change-Id: I4a36c3ffb621c03f7c355b4459d23f59071d9938