Update get_vm_power_state_listener to retrieve a poller
At the moment, the event polling is done by the caller, using the
listener we return here. This means that Hyper-V generated events
will 'leak' out of the library.

This patch adds an optional argument so that the caller can request
a handler instead, so it does not have to handle hyper-v specific
objects and WMI exceptions, which in fact should be done by os-win.
This event handler will invoke a callback with the VM name and power
state, as that's all the caller will care about.

Change-Id: Iddc52878bcc84bf3d8c31edbcd178eac4b59092d