Fix livemigrationutils wmi connection getter
The BaseUtilsVirt constructor overides the _wmi_namespace template,
expanding it with the host it retrieves.

This causes issues to the livemigrationutils connection getter
as this relies on the _wmi_namespace attribute to build it's remote
WMI connection. This is a regression caused by this change:
Iaa68a4ccd333b9e23cc7c373571f34d56ca2857b

This patch fixes the issue by simply avoiding overriding this
attribute.

Change-Id: Ifdba8fb1dcf9bf428c4da40103f224a2714c6586