Fix serial console worker blocking Nova
The worker used by Nova to log instance serial console output can
log an exception message.

The issue is that logging a message from a different thread causes
Nova to hang. It seems that the logging file handler causes issues
when greenthreads and multiple native threads are used at the same,
and the native threads log messages.

This patch removes this log message for now. In the future, we
should fix the logging handler to avoid this issue.

Closes-Bug: #1557498

Change-Id: Ia5c1aba28347648e90f5ba6111fc047d711236db