Ensure namedpipe IO workers clean up handles when stopping
At the moment, when an exception is encountered by the namedpipe
IO workers (for example, the named pipe is closed), they are
stopped but the file handles remain open until the namedpipe
handler is explicitly stopped.

This patch changes this, ensuring that when a named pipe IO worker
ends its execution, it gracefully closes the file handles.

Change-Id: If696b082fee719c6d3bb50888f87c1f0aeea5331