switch to post-versioning
This change removes the version entry from setup.cfg, relying on pbr to
derive the version number from the git tags. This post-versioning model
brings os-win in line with the other official OpenStack projects, and
avoids the need to submit separate patches to "bump" the version number
before and after a release.

Change-Id: I6d6976bc6cd1edd70864d9b2a1a8a3330b4bfcdb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>