Fix named pipe handler cleanup regression
This change introduced a regression related to stopping named pipe
IO workers: If696b082fee719c6d3bb50888f87c1f0aeea5331

Now that the IO workers close the named pipe when ending their
execution, attempting to cancel pending IO operations may raise
an invalid handle error. Simply checking if the pipe handle is
set was not enough, as we're open to a race condition.

It's safe to ignore this error, as long as we pass the overlapped
structures to the CancelIO function, ensuring that we don't have
issues in case the pipe was closed and the handle reused.

Also, it may happen that another IO request was issued right after
we've set the stopped event and canceled pending requests. In this
case, a retry will ensure that the IO workers are stopped properly
and that there are no more outstanding IO operations.

Closes-Bug: #1562473

Change-Id: Ib9ea3385daf9107346e85968e2292d0914c11a48