Consistently raise exception if port not found
Doesn't raise an exception on remove_switch_port. If a VM is
destroyed, all of its associated vNICs and switch ports are also
destroyed. The Neutron Hyper-V Agent will notice this fact and
call the remove_switch_port. The Agent will also receive a port_delete
RPC call for the already destroyed port.

Closes-Bug: #1563775

Change-Id: I991e3f664868266e84fde1471b38965b7ad04ca4