Properly handle stopping VM jobs
When destroying a VM, we attempt to terminate the jobs affecting it.

We just loop through the jobs and try to kill them, but by the
time we do this, the jobs can disappear, in which case we'll get a
"Not Found" exception.

In this case, it is safe to catch this exception and ignore it.

After looping through the jobs, we'll just fetch again the jobs
affecting the element, raising an exception only if there still
are pending jobs.

Also, if a job is killed, the check_ret_val method now raises an
exception instead of wrongfully reporting it as successfully
completed.

Conflicts:
	os_win/exceptions.py

Closes-Bug: 1566292

Change-Id: I93f9b5be11a7dc1b81b3d896fb88c0b0b4c4b31c
(cherry picked from commit bcc5ce23b6f378021642cc603517ae1eebe62092)