Fix retrieving VM physical disk mapping
When retrieving the VM 'physical' disks (host attached passthrough
disks) mapping, we attempt to retrieve the disk resource path, using
the 'Path_' method.

The issue is that when using the old WMI lib, this method is
case-insensitive, while when using PyMI, it's not.

This will raise a MI_METHOD_NOT_FOUND exception. This can easily be
fixed by using the right method name.

Change-Id: Iea2e22a67065f58a67ad1a89e6076408e5504be9
Closes-Bug: #1567916