Fix retrieving VM notes race condition
If the VM is deleted while we attempt to retrieve the VM notes, we
sometimes get a TypeError when we try to iterate over them, as the
Notes property will be set to None.

This change fixes the issue by storing this field in a different
variable, before attempting to use it.

Closes-Bug: #1569815

Change-Id: I89e73df9406a55e087751d20450e134be60867e6
(cherry picked from commit 70d3658d2635317bd785532c7d6fbc858ff05abe)