Adds missing attribute from get_cpus_info query
There is a new field added in cpu_info called MaxClockSpeed.
The field is not included in the query so its value will not
be retrieved with the query and it will be always None.

This patch adds the missing field in the query in order
to retrieve its value.

Closes-Bug: #1570345

Change-Id: Ie6d531ccd9fdb74c700dc6f52674d46838f5ee1a