iSCSI utils: accept rescan_attempts arg in login method
In order to be a drop in replacement for the old iSCSI utils module
and easily be used by the Nova driver, the new iSCSI initiator
utils module had to keep the same interface.

The login_storage_target method will call by default the
'ensure_lun_available' method, which will perform some rescans if
needed, after establishing the iSCSI session.

For this reason, it would make sense for this method to accept
the maximum number of rescans as an argument that will be passed
to 'ensure_lun_available'.

This will be helpful for the os-brick Windows iSCSI connector.

Change-Id: Ibbf134931a43dcf3e64f234582a8ebe079e65184