Add method for checking whether a share is local
Mounting loopback highly available shares is not supported.

The Hyper-V Nova driver needs a way to check whether a share is
local and use the local path in that case.

This change adds a method that attempts to resolve the share server
address and compare the IPs with the local IPs.

In case of a share exposed by a Scale-Out server, we'll use the
Distributed Node Name of the share, which points to all the nodes
of the cluster.

ddt is added as a requirement.

Change-Id: I04ac0b514ccac9c910728da205f0f0f11b43509a
Partial-Bug: #1580122