Add instance job terminate timeout
This change lets the caller set a timeout for the job terminate
operation. If the timeout is None or 0, this operation will wait
undefinetely for the jobs to be terminated (while attempting to
stop them).

Also, error state jobs are now properly handled, as we were not
considering them to be finished.

Change-Id: I47beec8b36cecd9db3960c2deff09e342418046a
(cherry picked from commit b5d60890e27b05aee7a23d4bbcc5fad69b2407c3)