Fix clustered vm live migration
At the moment, in order to migrate a clustered VM, we move the
according cluster resource group using the MoveToNewNodeParams WMI
method. This method accepts an undocumented 'Parameters' argument,
which is then passed to the underlying MoveClusterGroupEx clusapi.dll
function.

While the latter expects a property list as described at the following
link, we were passing the desired migration type directly.
https://msdn.microsoft.com/en-us/library/aa371809(v=vs.85).aspx

This worked on Windows Server 2012R2 but is causing issues on WS 2016.
Resource group end up hanging in a 'Pending' state.

This change addresses this issue by using clusapi.dll functions directly,
as the according WMI methods are not working, throwing a generic error
even though the VMs are migrated properly.

Closes-Bug: #1618425

Change-Id: Idfcd3505cbbf2754addeba4f1ebeb880f3b9a56b