Non-clustered VM live migration fix
When live migrating instances, we always request a
'VirtualSystemAndStorage' migration type.

In general, this works with shared storage as well, in which case
the instance disks should no be moved. Still, it's causing issues
when the instance disk files are placed on storage spaces disks.

This can be avoided by requesting the instance disks not to be
migrated.

This change addresses this issue, adding the 'migrate_disks' switch
argument to the 'live_migrate_vm' method.

Change-Id: I48dc6da29534bf0c477c742311ac15438d8b63ad
Closes-Bug: #1618446
(cherry picked from commit fcdbf99153ab48eeee2f641777b3fb10c06db02f)