Adds Hyper-V PCI Passthrough support
Discrete Device Assignment is a new feature in Windows Server 2016,
offering users the possibility of taking some of the PCI Express
devices in their systems and pass them through directly to a guest VM.

Adds a method in hostutils which retrieves the list of PCI devices
which can be be assigned to VMs.

Adds methods for attaching / detaching a PCI device to / from a VM
in vmutils.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Change-Id: I8e7782d3e1e9f8e92406604f05504a7754ffa3c2
Implements: blueprint os-win-pci-passthrough