Fix clustered VM migration status polling
When migrating a clustered VM, we rely on the cluster resource
group status in order to determine whether the VM was properly
migrated.

After a migration is requested, the resource group immediately
enters a 'pending' state. At the moment, the method polling the
resource group state uses a list of valid transition states,
incorrectly appending the desired state of the resource group
to this list.

This causes issues if the migration fails, as we're not going to
detect this and we'll continue polling indefinitely (unless a
timeout is passed).

This change fixes this issue by removing the desired state from
the valid transition states list.

Change-Id: Id1bdd6ccc6a2a6abc99e86ca362e03eb5adb66a2
Closes-Bug: #1628938
(cherry picked from commit d4ad19dc0fa90466d2b0ac0b9daa0830fdb6c467)