Ensure GetLastError gets called in the right thread
Recently, we've started invoking Win32 API functions in different
threads in order to avoid blocking other greenthreads.

The only issue is that if a function fails, we're retrieving the
error code by calling GetLastError from the original thread, thus
retrieving wrong error codes.

This change ensures that the subsequent error handling will be
performed in the same thread as the invoked function.

Change-Id: Id54b5bdb8e4561aaa42164b8c5d52a1a83da7149
Closes-Bug: #1632297