Allow setting VM snapshot types
On WS 2016, Hyper-V will attempt to take production checkpoints by
default, if the guest advertises this feature. The issue is that
the mechanism that automatically detects the VM checkpoint type
supported by the guest is not entirely safe.

We've had issues with older LIS versions in which case this feature
was not properly advertised and production checkpoints were attempted
but failed.

This change will allow explicitly setting the VM snapshot type to
be used for a given instance.

Partial-Bug: #1664941

Change-Id: I9fba30143bb0f653f1aaf7c97e5c3008a9f03120