Change intermediary vm state handling
os-win reports vms in intermediary power states as being running
(we're actualy stating this in an inline comment).

For some reason, one exception to this rule is the 'shutting down'
power state, for which we define a separate vm state.

The Nova Hyper-V driver then treats this speific state as a shutoff
state. This confuses the Nova manager in case of rebooting instances,
treating this as an unexpected instance shutdown, forcing a shutdown
afterwards.

While this should be fixed in the Nova driver as well, simply remaining
consistent with the above mentioned rule and reporting all intermediary
states as 'active' at the os-win level will mitigate the issue.

Closes-Bug: #1660978

Change-Id: I82bc176e17606684d4a88adaa99a171360e13e04