removes the VirtualSystemType kwarg when fetching VMs
A VM can be either "Realized" or "Planned", but not both. Thus,
using the VirtualSystemType as a filter is not useful.

Because of this, the "is_planned_vm" argument from vmutils
methods will no longer be unused, and will be removed once it is
no longer used.

This will make vmutils easier to use for Planned VMs, which will
be used for cold migration in the future. Plus, not all vmutils
methods have the "is_planned_vm" argument, which would have been
necessary in order to update the Planned VM (e.g.: get_vm_config_dir method).

Partial-Bug: #1663238

Change-Id: Ic153dcfd621a1e6235dd3a75241274ee7d7c868b