vmutils: set all *DataRoot paths to the same value
In order to maintain consistency with create_vm, update_vm
should also set the *DataRoot paths to the same given value.

Furthermore, this will be required for planned VMs, in order to
prevent having instances running on different compute nodes and
different storages to have some of the *DataRoot paths pointing to
the old compute node.

Partial-Bug: #1663238

Change-Id: Ic9248f44de6ff097b8080a80fd3212eeef6712f6