Fix share local address check
We're resolving share addresses to determine if those point to a
local address. The issue is that we fetch the local ips resolving
the local hostname, which won't include '127.0.0.1' or '::1'.

For this reason, a share such as '\\localhost\share' won't be
reported as being local.

This change handles this case.

Closes-Bug: #1680029

Change-Id: I36df4aa79bb7af548b6e39f9f9e7c4f979ae0a77