Allow aggregating disk rescan requests
Disk rescans are time consuming operations. We're trying to reduce
the amount of rescan calls by aggregating them. Basically, if a
rescan is already in proggress, we're waiting for it to complete
without issuing a new request.

This behavior is disabled by default and can be enabled using the
'merge_requests' argument.

Change-Id: I6132d05c6ca7725226348f9fa8237e1810e65a7d