raises exceptions.NotFound instead of x_wmi: Not Found
When an instance is being rebuilt, or shelved and unshelved, the VM gets
destroyed and recreated, keeping the same neutron port. If a reference to
a non-existent object is passed to a WMI method (or no longer exists), an
x_wmi: Not Found exception is raised.

Adds decorators to the jobutils methods, which will reraise
os_win.exceptions.NotFound exceptions instead of the x_wmi: Not found
exceptions.

Change-Id: Ic371cd1e2c2a54c4f386c02557bc9cdd9518bc9b
Closes-Bug: #1694432