Fixes cached old WMI service objects issue
When using WMI, the resources passed to the
Msvm_VirtualSystemManagementService are not cleaned up
by the garbage collector if the Service object is still
referenced (cached as a property in BaseUtilsVirt).

PyMI doesn't have this issue, so the Service objects can be
safely cached.

This issue only affects Windows / Hyper-V Server 2012.

Conflicts:
        os_win/tests/unit/utils/compute/test_migrationutils.py

Closes-Bug: #1697389
(cherry picked from commit 49a012e83c44d28ee533aace22c42b1781b8e5e2)

tests: Sets baseutils.BaseUtilsVirt._old_wmi

There is a race condition between tests, and because
the _old_wmi is not set, it will try to load it, but since it
is mocked, it cannot, and it shouldn't.

Change-Id: Id029baf2e72f7b67eb6af3b3e66f998b902fd412

Change-Id: If8df0bf0a2f13a9e94993dde8b2a18943f2e9f83