Improve WMI job check method
At the moment, we're raising a HyperVException when a WMI job fails.
The issue is that this doesn't include the job status or error code
as attributes, which is troublesome when we need to have exception
handling logic relying on that.

This change adds a new exception that will include those fields,
also simplifying the WMI job status check method.

Note that although WMI jobs are not necessarily Hyper-V related,
we're inheriting HyperVException as that's what Nova expects at the
moment.

Also, we're now following the same logic as the HyperV PowerShell
module in terms of checking a WMI job state. States such as
"CompletedWithWarnings" (32768) will now be gracefully handled,
instead of throwing an exception.

Another thing to note is that we're now always treating the
'Exception' job state as a final state.

Partial-Bug: #1694661

Change-Id: I498281d3af673e4e6ac3adc2fb1a566efaba9d4e