Wrap "WMI not found" exception
There are situations in which a VM unexpectedly disappears while
we're trying to perform an operation on it. Raising an appropriate
os-win exception in such cases helps os-win consumers perform
appropriate error handling.

This change adapts the decorator translating such exceptions,
using it for the 'get_vm_summary_info' method. This will help Nova
properly handle the case in which an instance is deleted while an
instance lifecycle event is checked.

Related-Bug: #1653862

Change-Id: I6930ed9df9a79e4f8f01bb96ce6dd942fd4760f8