iSCSI: add checks for MPIO disks
This change adds an extra parameter to the method which ensures
that a lun is available. This allows verifying that the according
disk was claimed by the MPIO service.

Note that for convenience reasons, the 'ensure_lun_available'
method arguments are now exposed by the 'get_device_number_and_path'
method.

Change-Id: If5f00a019be76c09724f4e93f3d53431286d0efa
Related-Bug: #1694671