FC: cleanup WWN usage
At the moment, we're using WWNs represented either as byte arrays,
hex strings or ctypes structures. Needless to say, this is quite
confusing as the fc_utils methods accept multiple formats.

As per this change, we're accepting WWNs represented as hex strings
in case of the public methods (just as before), while for private
methods we're using ctypes WWN structures (containing 8 byte arrays),
which in turn are passed to hbaapi.dll functions.

Some arguments are renamed/documented accordingly. Note that the
helper methods converting those WWNs are moved to the '_utils'
module, as those will be reused by a subsequent change.

Change-Id: I43daf4ad943d0097c951c2ad4ccb856a2fe9e1bc