vmutils: Fixes setting disk QoS on Hyper-V 6.2
Hyper-V disk QoS is supported on Windows / Hyper-V Server 2012 R2
(OS version 6.3) or newer. It cannot be set on earlier versions.

If the set_disk_qos_specs method is called on Windows / Hyper-V Server
2012 with non-zero values, an os_win.exceptions.UnsupportedOperation
exception will be raised.

If the given min_iops and max_iops arguments are None or zero, the
method will be no-op: QoS doesn't need to be enabled, nor disabled
(0 means that QoS is disabled).

Change-Id: Ia3b47477ba6814153bf0ab74ec9737a01bb11678
Closes-Bug: #1716891