Fix fetching vm boot sources
When setting the VM disk boot order, we're passing Msvm_DiskDrive
WMI object paths in case of passthrough disks and the actual image
file paths in case of vhd/x images.

In order to determine the disk type, os-win checks if the path is
an existing file path (using os.path.exist), which is extremely
inefficient. The call can take up to a minute in case of a
passthrough disk.

This patch updates the logic so that we're checking if the provided
path resembles a Msvm_DiskDrive WMI object path. The
'_is_drive_physical' method is removed due to its limited scope.

(cherry-picked from commit 7af8caeb5e79197656133ad28368e1012227675d)

Change-Id: I0fc7a8b5f1feabe07fc1707d8469b929b3739a1d
Closes-Bug: #1723448