Add method returning vswitch extensions
This change adds a new method returning available extensions for
a given vswitch. It returns a list of dicts that include vswitch
extension information.

This method is required by the Neutron Hyper-V agent in order to
validate a configured vSwitch.

Note that this change also adds the CIM enabled state constants
(which are re-used for various type of resources).

Related-Bug: #1592777

Change-Id: I0856a23030480d435fd0582fc6b4bc688c98ae93