[Trivial] Allow flushing named pipe log handles
The named pipe handler provided by os-win is used by Nova when
consuming instance serial ports. For convenience, it allows logging
the named pipe output to a file (useful for storing the serial
console output).

The issue is that the log file never gets flushed, thus Nova returns
incomplete results. This change adds a method that may be used for
flushing the log file.

Change-Id: I4b5eb7579ac0c92e9c3630539b0c4d76f5b98c6d
Closes-Bug: #1741494