Fixes port security settings caching issue
When updating the MAC spoofing for a given vswitch port, its
Msvm_EthernetSwitchPortSecuritySettingData object is removed
and a new one is created, but the original object is not
cleared from the cache. Because of this, the next update will
fail with a Not Found exception.

This patch addresses this issue.

Closes-Bug: #1744755

Change-Id: I91b9000c64d269b7d5b1db9e0c6344bbd70e37b2
(cherry picked from commit 776270c4b3be0523f2ee2a73792b4a41bb52d6f5)