Fix incorrect numeric comparisons
In a few places, we're using "is" or "is not", when comparing
numeric variables.

This is incorrect and will give unexpected results when the types
do not match (for example, if one of the variables is a long int):
http://paste.openstack.org/raw/730618/

This change fixes those comparisons.

Closes-Bug: #1794068

Change-Id: Ibf28778879050cfdfa82ef4f9b587170dde6203f
(cherry picked from commit 1f8200ef7909b13adf501f2805554160fa934f3a)