Fix port metrics
At the moment, port metrics are not enabled. The reason is that we're
using the wrong switch port WMI class.

This issue went more or less unnoticed, especially due to the fact that
we're currently ignoring the according WMI method return code.

This change ensures that port metrics are properly enabled and that an
exception will be raised when failing to do so.

Closes-Bug: #1794291

Change-Id: I8086e7dd9e134749760b05469f5b2c8bcfdcdbde