Cleanup utilsfactory map
This change cleans up the utilsfactory map, making it easier to add
new classes.

Most classes have 6.2 as a minimum version, without having a maximum
version cap. For this reason, this change sets those versions as
defaults so that we don't have to be explicit every time.

At the same time, we're now using a list for each class type, instead
of a dict that uses the class name as a key (which is redundant). So,
something like this:

'networkutils': {
    'NetworkUtils': {
        'min_version': 6.2,
        'max_version': 6.3,
        'path': 'os_win.utils.network.networkutils.NetworkUtils'},
    'NetworkUtilsR2': {
        'min_version': 6.3,
        'max_version': None,
        'path': 'os_win.utils.network.networkutils.NetworkUtilsR2'}},

... becomes

'networkutils': [
    {'max_version': 6.3,
     'path': 'os_win.utils.network.networkutils.NetworkUtils'},
    {'min_version': 6.3,
     'path': 'os_win.utils.network.networkutils.NetworkUtilsR2'},
],

Change-Id: I7919617166b2e872f1b50c0a0f40196fc90e9a82