Fix VHD detach
Attempting to disconnect a VHD from the local host may fail if the
image is in-use by a different host, even though it's not actually
mounted on the local host.

The reason is that the detach API requires a handle, which we may
fail to obtain if the file is in-use.

This change will try to figure out if the image is actually attached
to the local host before erroring out. This check is a bit more
complicated than it should, again due to file locks. We'll use
two approaches:
* the preffered way, using virtdisk.dll:GetVirtualDiskInformation,
  which may fail due to file locks
* the slower way, used as a fallback, which involves getting a full
  list of attached images and comparing file and volume ids as
  the file may be accessible through multiple paths

Closes-Bug: #1840954
Change-Id: Ib5bf0070c0ba4cb18e98c1cb983c641eaae582bf