Add configurable timeout for in-use files
In some cases, Hyper-V locks on VHD images persist for a few seconds
even after the instances are deleted (not sure if that's due to
Hyper-V or the filesystem). In order to mitigate this, we're
already performing some retries when deleting instance files.

We've encountered similar issues with 3rd party backup tools
that kept open file handles.

This change adds a configurable timeout that will be used when waiting
for in-use fles. The default value is 15 seconds, slightly more than
the current hardcoded value. At the same time, this timeout will
also be used when deleting individual files or renaming folders or
files.

Change-Id: I49307146b79a20089190e1ce8bde73d656d37c6d