Avoid propagating wmi exceptions for soft shutdown
The soft shutdown operation is expected to fail while the VM is
still booting. The issue is that in some cases, os-win will
propagate a wmi exception that Nova cannot handle:

http://paste.openstack.org/raw/800833/

That's because of Msvm_ShutdownComponent.InitiateShutdown, an
asynchronous WMI method that can sometimes throw a wmi exception
right away.

For now, we're taking care of this specific WMI call only. If needed,
we may consider handling all async WMI methods by introducing a new
decorator or helper function.

Change-Id: I9263b14c3c343db6453645152cb9274e38864b48
Closes-Bug: #1907270